Fix issues related to removing the currently playing tl_track:
Currently, removing the current tl_track from the tracklist does not cause playback to stop.
This means GStreamer keeps playing it and we have no reference to it, so we can't do stop() or next() or prev().
The only way to clear this state is to start a new track playing.

Also currently, stopping playback with consume enabled does not remove the current tl_track from the tracklist until playback starts again.
This leads to the same issue as above where a track is playing that is not in the tracklist.

This change does the following:
Ensure we emit track_playback_ended when changing playback state to STOPPED.
This makes sure the track gets consumed - both on stop and on end_of_stream.

When removing a track from the tracklist, if it is the current tl_track then stop playback first
BUT don't do this if the remove is being called to consume a track, as this makes playback stop after every track and results in recursive calls to remove().
This is achieved with a new private method which has a flag to control this behaviour

Also when investigating this I discovered multiple race conditions that need to fixed because this
change makes them easier ot provoke. There is a whole battery of new tests that aims to cover all the possible
race conditions.